% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{alphaDisconfirmation}
\alias{alphaDisconfirmation}
\title{Calculates the intended probability of the hypothesis after a nonsignificant test.}
\usage{
alphaDisconfirmation(
  n,
  effsize,
  prior,
  postfalse,
  test = "t-test",
  type = "two.sample",
  alternative = "two.sided",
  k = 4
)
}
\arguments{
\item{n}{Number of participants per group.}

\item{effsize}{Effect size corresponding to statistical test. D for t-test, r for correlation and f for anova.}

\item{prior}{Prior probability of H1 being true.}

\item{postfalse}{The posterior probability that is aimed for if the test is not significant (should be higher than prior).}

\item{test}{Kind of statistical test. Can be "t-test", "anova" and "correlation".}

\item{type}{Indicate whether it is a one.sample, or two.sample test (for t-test only).}

\item{alternative}{Indicate whether the alternative is "two.sided", "less", or "greater".}

\item{k}{Number of groups for anova.}
}
\value{
A data frame with the alpha level, the  power, the expected correct change in belief in percent and posterior probability after a significant result.
}
\description{
The user specifies the intended probability of the hypothesis after a non-significant test result and the alpha level is adjusted accordingly.
}
\examples{
alphaDisconfirmation(100, 0.3, 0.5, 0.2, test = "correlation").

}
