% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{idealalpha}
\alias{idealalpha}
\title{Idealalpha}
\usage{
idealalpha(
  n,
  effsize,
  prior,
  test = "t-test",
  type = "two.sample",
  alternative = "two.sided",
  k = 4
)
}
\arguments{
\item{n}{Number of participants per group.}

\item{effsize}{Effect size corresponding to statistical test. D for t-test, r for correlation and f for anova.}

\item{prior}{Prior probability of H1 being true.}

\item{test}{Kind of statistical test. Can be "t-test", "anova" and "correlation".}

\item{type}{Indicate whether it is a "one.sample", or "two.sample" test (for t-test only)}

\item{alternative}{Indicate whether the alternative is "two.sided", "less" or "greater"}

\item{k}{Number of groups for anova.}
}
\value{
A data frame with the ideal alpha level, the ideal power, the expected correct change in belief in percent and the
expected change in belief under alpha = 0.05.
}
\description{
Calculates the alpha level and power which result in the highest informational value of a study.
}
\examples{
idealalpha(100, 0.3, 0.5, test = "correlation")

}
