% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{probpower}
\alias{probpower}
\title{Probability Based Power Analysis}
\usage{
probpower(
  effsize,
  prior,
  postfalse,
  posttrue,
  test = "t-test",
  type = "two.sample",
  alternative = "two.sided",
  k = 4
)
}
\arguments{
\item{effsize}{Effect size corresponding to statistical test. D for t-test, r for correlation and f for anova.}

\item{prior}{Prior probability of H1 being true.}

\item{postfalse}{The posterior probability that is aimed for if the test is not significant (should be higher than prior).}

\item{posttrue}{The posterior probability that is aimed for if the test is significant.}

\item{test}{Kind of statistical test. Can be "t-test", "anova" and "correlation".}

\item{type}{Indicate whether it is a "one.sample", or "two.sample" test (for t-test only).}

\item{alternative}{Indicate whether the alternative is "two.sided", "less", or "greater".}

\item{k}{Number of groups for anova.}
}
\value{
A data frame with the alpha level, the  power, the expected probability after a significant result and a non-significant result, the expected amount of learning, and needed sample size.
}
\description{
Set both the posterior probability after a positive result and the posterior probability after a negative result
The function will calculate the required sample size and alplha level to achieve this.
}
\examples{
probpower(0.3, 0.5, 0.2, 0.8, test = "correlation").

}
